+++ src/main/resources/META-INF/MANIFEST.MF
@@ -1,4 +1,3 @@
 Manifest-Version: 1.0
-Class-Path: /
-Main-Class: MapMatching
-
+Class-Path: geo
+Main-Class: geo.algorithms.App
