+++ src/main/scala/geo/data/Write.scala
@@ -1,0 +1,5 @@
+package geo.data
+
+object Write {
+
+}
