+++ src/main/scala/geo/elements/Way.scala
@@ -1,0 +1,5 @@
+package geo.elements
+
+class Way (val points: List[Point]) extends Serializable{
+
+}
